// frontend/src/services/OpenApi/models/CommentDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CodeDescriptionDto } from './CodeDescriptionDto';
export type CommentDto = {
    commentSource: CodeDescriptionDto;
    commentType: CodeDescriptionDto;
    commentText: string | null;
};